requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
click==7.1.2
discord.py==1.4.1
Flask==1.1.2
idna==2.10
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
pyee==7.0.3
python-dotenv==0.14.0
slackclient==2.8.2
slackeventsapi==2.2.1
typing-extensions==3.7.4.3
wakeonlan==1.1.6
Werkzeug==1.0.1
yarl==1.5.1
